backend/requirements.txt
fastapi
uvicorn[standard]
celery
redis
httpx
requests
loguru
python-dotenv
